+++ macros/src/derive.rs
@@ -1,4 +1,5 @@
 mod deserialize;
+mod deserialize_key;
 mod introspectable;
 mod options;
 mod serialize;
@@ -7,6 +8,7 @@ mod serialize_key;
 use options::{ItemOptions, Options};
 
 pub use deserialize::{gen_deserialize_from_aldrin, gen_deserialize_from_core};
+pub use deserialize_key::{gen_deserialize_key_from_aldrin, gen_deserialize_key_from_core};
 pub use introspectable::{gen_introspectable_from_aldrin, gen_introspectable_from_core};
 pub use serialize::{gen_serialize_from_aldrin, gen_serialize_from_core};
 pub use serialize_key::{gen_serialize_key_from_aldrin, gen_serialize_key_from_core};